feat(WIP-shipment): add basic shipment functionality
